requirements.txt
python-dotenv==0.13.0
paho-mqtt==1.5.0
